+++ Source/Csla.Generators/cs/Csla.Generator.AutoSerialization.CSharp.TestObjects/AddressPOCO.cs
@@ -1,0 +1,28 @@
+﻿//-----------------------------------------------------------------------
+// <copyright file="AddressPOCO.cs" company="Marimer LLC">
+//     Copyright (c) Marimer LLC. All rights reserved.
+//     Website: https://cslanet.com
+// </copyright>
+// <summary>Class that can be used for testing serialization behaviour</summary>
+//-----------------------------------------------------------------------
+using Csla.Serialization;
+
+namespace Csla.Generator.AutoSerialization.CSharp.TestObjects
+{
+
+  [AutoSerializable]
+  public partial class AddressPOCO
+  {
+
+    public string? AddressLine1 { get; set; }
+
+    public string AddressLine2 { get; set; } = string.Empty;
+
+    public string Town { get; set; } = string.Empty;
+
+    public string County { get; set; } = string.Empty;
+
+    public string Postcode { get; set; } = string.Empty;
+
+  }
+}